Crate dinic

Solves maximum flow problem.

Basic usage

First, initialize Dinic with the number of vertices.

Insert edges with a method add_edge.

Execute the algorithm by calling a method flow with the source s and sink t.

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);

let flow = dinic.flow(0, 2);
assert_eq!(flow, 30);

Restore the minimum cut

If flow has called exactly once before, min_cut will return the minimum cut. See the API document to detailed specs.

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);
dinic.flow(0, 2);

assert_eq!(dinic.min_cut(0).as_slice(), &[true, false, false]);

Get the state of an edge or a vertex

You can query the state of an edge via get_edge. An EdgeKey object returned by add_edge is necessary to query it.

use dinic::Dinic;
use dinic::Edge;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
let key = dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);
dinic.flow(0, 2);

assert_eq!(dinic.get_edge(key), Edge {
    from: 1,
    to: 2,
    cap: 15,
    flow: 10
});

Moreover get_edges, get_network will summarize the whole network.

use dinic::Dinic;
use dinic::Edge;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
let key = dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);
dinic.flow(0, 2);

let edges = dinic.get_edges();
let network = dinic.get_network();

// 0th edge
assert_eq!(network[0][0], edges[0]);
assert_eq!(edges[0], Edge {
    from: 0,
    to: 1,
    cap: 10,
    flow: 10
});

// 1st edge
assert_eq!(network[1][0], edges[1]);
assert_eq!(edges[1], Edge {
    from: 1,
    to: 2,
    cap: 15,
    flow: 10
});

// 2nd edge
assert_eq!(network[0][1], edges[2]);
assert_eq!(edges[2], Edge {
    from: 0,
    to: 2,
    cap: 20,
    flow: 20
});

You also can get the excess of vertices, but we do not provide an interface to get the excess of a vertex because it will take O ( m ) time.

use dinic::Dinic;
use dinic::Edge;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
let key = dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);
dinic.flow(0, 2);

assert_eq!(dinic.get_excess().as_slice(), &[-30, 0, 30]);

If you use an unsigned type, get_excess will surely overflow because the excess of the source is almost always negative.

use dinic::Dinic;
use dinic::Edge;

let mut dinic = Dinic::<u32>::new(2); // Force to use `u32` instead of `i32`.
dinic.add_edge(0, 1, 10);
dinic.flow(0, 1);

dinic.get_excess(); // panics

Call flow more than once

You can call flow more than once. flow(s, t) will augment the flow from s to t as much as possible. If flow is called with different s or t from the previous ones, it may yield non-zero excess at more than two points.

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);

dinic.flow(0, 2);
let aug = dinic.flow(1, 2);
assert_eq!(aug, 5);
assert_eq!(dinic.get_excess().as_slice(), &[-30, -5, 35]);

Change the capacity or the amount of flow. (dangerous operation)

change_edge changes the capacity and the amount of flow of an edge. While it is so dangerous operation (safe variant wanted!), it is sometimes useful, for example, when solve a maximum flow problem with lower limit. See the API document for detailed specs.

Structs

Dinic

A struct to execute Dinic’s algorithm.

Edge

A summary of the state of an edge, which is returned by Dinic::get_edge.

EdgeKey

A key object to query an edge.

Traits

Value

An adapter trait of the capacity.