Struct dinic::Dinic

pub struct Dinic<T> { /* private fields */ }

A struct to execute Dinic’s algorithm.

See the module level documentation.

Implementations

impl<T> Dinic<T>where T: Value,

pub fn new(n: usize) -> Self

Creates a new instance of Dinic

Examples

use dinic::Dinic;

let mut dinic = Dinic::<u32>::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);

let flow = dinic.flow(0, 2);
assert_eq!(flow, 30);

pub fn add_edge(&mut self, from: usize, to: usize, cap: T) -> EdgeKey

Inserts a new edge to the network.

Constraints

from, to < n

T::zero() <= cap

Complexity

O ( 1 ) amortized.

Examples

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);

let flow = dinic.flow(0, 2);
assert_eq!(flow, 30);

pub fn flow(&mut self, s: usize, t: usize) -> T

Auguments the flow from s to t as much as possible. It returns the amount of the flow augmented.

You may call it multiple times. See the module level documentation.

Constraints

s != t,

The answer should be in T.

Complexity

O ( min ( n^{2/3} m, m^{3/2} ) ) if all the capacities are 1

O ( n^2 m )

Examples

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);

let flow = dinic.flow(0, 2);
assert_eq!(flow, 30);

pub fn flow_with_limit(&mut self, s: usize, t: usize, flow_with_limit: T) -> T

Auguments the flow from s to t as much as possible as long as not exceeding flow_limit. It returns the amount of the flow augmented.

You may call it multiple times. See the module level documentation.

Constraints

s != t,

The answer should be in T.

Complexity

O ( min ( n^{2/3} m, m^{3/2} ) ) if all the capacities are 1

O ( n^2 m )

Examples

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);

let flow = dinic.flow_with_limit(0, 2, 28);
assert_eq!(flow, 28);

pub fn min_cut(&self, s: usize) -> Vec<bool>

Returns a vector of length n, such that the i-th element is true if and only if there is a directed path from s to i in the residual network. The returned vector correponds to a s -- t minimum cut after calling self.flow(s, t) exactly once.

Examples

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);

dinic.flow(0, 2);
assert_eq!(dinic.min_cut(0).as_slice(), &[true, false, false]);

pub fn get_edge(&self, edge_key: EdgeKey) -> Edge<T>

Returns the current internal state of the edges.

Complexity

O ( 1 )

Examples

use dinic::Dinic;

let mut dinic = Dinic::new(3);
let edge_0 = dinic.add_edge(0, 1, 10);
let edge_1 = dinic.add_edge(1, 2, 15);
let edge_2 = dinic.add_edge(0, 2, 20);

let flow = dinic.flow(0, 2);

assert_eq!(dinic.get_edge(edge_0).flow, 10);
assert_eq!(dinic.get_edge(edge_1).flow, 10);
assert_eq!(dinic.get_edge(edge_2).flow, 20);
assert_eq!(flow, 30);

pub fn get_edges(&self) -> Vec<Edge<T>>

Collects all the edges.

Edges are sorted in order of addition.

Complexity

O ( m )

Examples

use dinic::Dinic;
use dinic::Edge;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);
dinic.flow(0, 2);

let edges = dinic.get_edges();
assert_eq!(edges[0], Edge {
    from: 0,
    to: 1,
    cap: 10,
    flow: 10
});
assert_eq!(edges[1], Edge {
    from: 1,
    to: 2,
    cap: 15,
    flow: 10
});
assert_eq!(edges[2], Edge {
    from: 0,
    to: 2,
    cap: 20,
    flow: 20
});

pub fn get_network(&self) -> Vec<Vec<Edge<T>>>

Collects all the edges and arrange it in adjacent-list style.

In each row, edges are sorted in order of addition.

Complexity

O ( n + m )

Examples

use dinic::Dinic;
use dinic::Edge;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);
dinic.flow(0, 2);

let network = dinic.get_network();
assert_eq!(network[0][0], Edge {
    from: 0,
    to: 1,
    cap: 10,
    flow: 10
});
assert_eq!(network[0][1], Edge {
    from: 0,
    to: 2,
    cap: 20,
    flow: 20
});
assert_eq!(network[1][0], Edge {
    from: 1,
    to: 2,
    cap: 15,
    flow: 10
});

pub fn get_excess(&self) -> Vec<T>

Returens the Vec of excess of all the vertices.

i-th entry is the excess of vertex i.

Complexity

O ( n + m )

Examples

use dinic::Dinic;

let mut dinic = Dinic::new(3);
dinic.add_edge(0, 1, 10);
dinic.add_edge(1, 2, 15);
dinic.add_edge(0, 2, 20);
dinic.flow(0, 2);

let excess = dinic.get_excess();
assert_eq!(excess.as_slice(), &[-30, 0, 30]);

pub fn change_edge(&mut self, edge_key: EdgeKey, new_cap: T, new_flow: T)

Changes the capacity and the amount of the edge corresponding to edge_key to new_cap and new_flow, respectively. It does not change the capacity or the flow amount of other edges. See the module level documentation.

Constraints

T::zero() <= new_flow <= new_cap

Complexity

O ( 1 )

Examples

Let us consider a bipartite matching problem on the graph K _ 2. Obviously this graph has a perfect matching.

o--o
 \/
 /\
o--o

It can be solved as a maximum flow problem.

use dinic::Dinic;

let mut dinic = Dinic::new(6);
let edge_0 = dinic.add_edge(0, 1, 1);
let edge_1 = dinic.add_edge(0, 2, 1);
let edge_2 = dinic.add_edge(1, 3, 1);
let edge_3 = dinic.add_edge(1, 4, 1);
let edge_4 = dinic.add_edge(2, 3, 1);
let edge_5 = dinic.add_edge(2, 4, 1);
let edge_6 = dinic.add_edge(3, 5, 1);
let edge_7 = dinic.add_edge(4, 5, 1);

let flow = dinic.flow(0, 5);
assert_eq!(flow, 2);

Now, let us add another constraint. Forbid to match the top ones. In order to do this, we deminish the flow by 1 from the source to the sink, along this matching edge, and diminish the capacity of this matching edge to 0, so that the maching is of size 1 and the dinic is also a feasible flow of value 1.

Now call Dinic::flow again with the same source and the same sink. So new matching will be of size 2 shaping like “X”. So Dinic::flow will return the delta 1 = 2 - 1 of the flow.

dinic.change_edge(edge_0, 1, 0); // the edge from the source
dinic.change_edge(edge_2, 0, 0); // the matching edge
dinic.change_edge(edge_6, 1, 0); // the edge to the sink

// now, `dinic` has a feasible flow of value `1`

let augment = dinic.flow(0, 5);
assert_eq!(augment, 1); // and augmented by `1` and became `2`.